finall update

+++ database/migrations/2022_10_07_073211_create_locations.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('email');
             $table->string('address');
             $table->bigInteger('business_id')->unsigned();
-            $table->bigInteger('user_id')->unsigned();   
+            $table->integer('user_id')->unsigned();   
             $table->timestamps();
             $table->foreign('business_id')->references('id')->on('businesses')->onDelete('cascade');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
